# openapi.yaml
openapi: 3.1.1
info:
  title: fb-marketing-server API
  version: 1.0.0
  summary: Private OpenClaw-to-Meta Ads service contract
  description: |
    Business API for explicitly scoped Meta Ads reads and human-approved mutations.
    OpenClaw is the sole caller. Meta credentials, arbitrary Graph forwarding, delete,
    public renewal, Google Drive, remote URL, and filesystem-path APIs are excluded.
    Meta Marketing API v26.0 is pinned for MVP; upgrades are planned and validated.
    Every Meta call appends redacted audit metadata with its request correlation ID.
servers:
  - url: http://127.0.0.1:3000
    description: Loopback-only local service
...
paths:
...
  /v1/capabilities:
    get:
      tags: [Discovery]
      summary: Discover safe capabilities for one client and Ad Account
      description: |
        Returns static supported campaign kinds plus currently accessible dynamic
        Meta assets and actionable capability gaps for the resolved scope. Dynamic
        IDs are opaque and are revalidated at proposal creation and execution.
        Assets are deterministically paginated; cursors bind the authenticated caller,
        client/account scope, and a hash of the optional asset-type filter.
        This endpoint performs no mutation, proposal creation, approval, or execution.
        Diagnosed Meta configuration and asset-access gaps return 200; 403 is only
        for local caller authorization failure. Any Meta validation call and its
        normalized outcome are audited with the request correlation ID.
      operationId: getCapabilities
      security:
        - ServiceBearer: []
      x-requirements:
        [FR-1, FR-2, FR-3, FR-8, SEC-3, SEC-7, SEC-8, NFR-4, NFR-8]
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientIdQuery"
        - $ref: "#/components/parameters/AdAccountIdQuery"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/CapabilityAssetType"
      responses:
        "200":
          description: Safe discovery metadata and capability status for the resolved scope.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapabilitiesResponse"
        "400":
          $ref: "#/components/responses/CapabilityBadRequest"
        "401":
          $ref: "#/components/responses/CapabilityUnauthorized"
        "403":
          $ref: "#/components/responses/CapabilityForbidden"
        "409":
          $ref: "#/components/responses/CapabilityScopeConflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          $ref: "#/components/responses/UpstreamError"
...
  /v1/media:
    post:
      tags: [Media]
      summary: Stage a trusted OpenClaw chat attachment
      description: |
        Accepts multipart binary only, never a local path or URL. The service verifies
        actual content type and limits, hashes bytes, and stages privately. Media is
        scope-bound and removed no later than 12 hours after staging or when its linked
        operation completes or expires, whichever is earlier.
      operationId: stageMedia
      x-requirements: [FR-1, FR-3, FR-6, SEC-3]
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MediaUpload"
            encoding:
              file:
                contentType: image/jpeg, image/png, video/mp4
              attachment:
                contentType: application/json
      responses:
        "201":
          description: Attachment validated and staged.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/ScopeConflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/UnprocessableMedia"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    ServiceBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-service-token
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      schema:
        $ref: "#/components/schemas/RequestId"
    ClientIdQuery:
      name: client_id
      in: query
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
    AdAccountIdQuery:
      name: ad_account_id
      in: query
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
...
    Cursor:
      name: cursor
      in: query
      description: |
        Opaque cursor bound to authenticated caller, resolved scope when applicable,
        filters, and request hash. Mismatched reuse returns 400.
      schema:
        type: string
        minLength: 1
        maxLength: 1024
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    CapabilityAssetType:
      name: asset_type
      in: query
      description: Optional asset discriminator. Its exact value is included in the cursor filter hash.
      schema:
        $ref: "#/components/schemas/CapabilityAssetType"
...
  headers:
    RequestId:
      required: true
      schema:
        $ref: "#/components/schemas/RequestId"
  responses:
    CapabilityBadRequest:
      description: |
        Invalid capability query or cursor binding. `resolved_scope` is forbidden.
        Each supplied query ID is copied to `supplied_scope` exactly as received,
        without normalization, substitution, or resolved labels; omitted IDs are not invented.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/CapabilityProblem400"
    CapabilityUnauthorized:
      description: |
        Service bearer is missing or invalid. `resolved_scope` is forbidden. Any
        supplied query IDs are preserved exactly without disclosing resolved labels.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/CapabilityProblem401"
    CapabilityForbidden:
      description: |
        Authenticated local caller lacks authority. `resolved_scope` is forbidden. Any
        supplied query IDs are preserved exactly without disclosing resolved labels.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/CapabilityProblem403"
    CapabilityScopeConflict:
      description: |
        Supplied client and Ad Account do not resolve to the same authorized scope.
        `resolved_scope` is forbidden and `supplied_scope` preserves both required
        query values exactly, without normalization, substitution, or resolved labels.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/CapabilityProblem409"
    BadRequest:
      description: Invalid input, scope pair, or cursor binding.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem400"
    Unauthorized:
      description: Service bearer is missing or invalid.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem401"
    Forbidden:
      description: Authenticated local caller lacks authority.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem403"
    NotFound:
      description: Supplied resource was not found or visible in its scope.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem404"
    ScopeConflict:
      description: Supplied client and Ad Account do not resolve to the same authorized scope.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem409Scope"
...
    PayloadTooLarge:
      description: Attachment exceeds an applicable limit.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem413"
...
    UnprocessableMedia:
      description: Media is unsupported, mismatched, expired, wrong-scope, or hash-invalid.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem422Media"
...
    RateLimited:
      description: Local or Meta request rate exceeded.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        Retry-After:
          schema:
            type: string
            pattern: "^(0|[1-9][0-9]*)$"
          description: Nonnegative delay in whole seconds serialized as an HTTP header string.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem429"
    UpstreamError:
      description: Normalized Meta upstream failure; raw bodies and secrets are excluded.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem502"
    InternalError:
      description: A normalized internal service failure with no secret or implementation detail.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem500"
  schemas:
    Identifier:
      type: string
      minLength: 1
      maxLength: 255
    RequestId:
      type: string
      minLength: 8
      maxLength: 128
      pattern: "^[A-Za-z0-9._:-]+$"
...
    Currency:
      type: string
      pattern: "^[A-Z]{3}$"
...
    ResolvedScope:
      type: object
      additionalProperties: false
      required: [client_id, client_name, ad_account_id, ad_account_name]
      properties:
        client_id:
          $ref: "#/components/schemas/Identifier"
        client_name:
          type: string
          minLength: 1
        ad_account_id:
          $ref: "#/components/schemas/Identifier"
        ad_account_name:
          type: string
          minLength: 1
        currency:
          $ref: "#/components/schemas/Currency"
        timezone:
          type: string
          description: IANA timezone reported by Meta when available.
    PageInfo:
      type: object
      additionalProperties: false
      required: [next_cursor]
      properties:
        next_cursor:
          type: [string, "null"]
...
    IntegrationState:
      type: string
      enum:
        [
          registered,
          validating,
          active,
          configuration_required,
          reauthorization_required,
          asset_access_required,
        ]
    Diagnostic:
      type: object
      additionalProperties: false
      required: [code, message, remediation]
      properties:
        code:
          type: string
          pattern: "^[a-z][a-z0-9_]*$"
        message:
          type: string
          minLength: 1
        remediation:
          type: string
          minLength: 1
    MetaPermission:
      type: string
      pattern: "^[a-z][a-z0-9_]{1,127}$"
      description: Safe Meta permission name, including newly observed permissions; never a token, secret, credential, or authorization value.
    IntegrationStatus:
      type: object
      additionalProperties: false
      required:
        [
          integration_id,
          generation,
          state,
          checked_at,
          missing_permissions,
          diagnostics,
        ]
      properties:
        integration_id:
          $ref: "#/components/schemas/Identifier"
        generation:
          $ref: "#/components/schemas/Identifier"
        state:
          $ref: "#/components/schemas/IntegrationState"
        checked_at:
          type: string
          format: date-time
        missing_permissions:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/MetaPermission"
        diagnostics:
          type: array
          items:
            $ref: "#/components/schemas/Diagnostic"
...
    CampaignKind:
      type: string
      enum: [SALES_WEBSITE, LEADS_WEBSITE, LEADS_INSTANT_FORM]
...
    CapabilityMatrix:
      type: object
      additionalProperties: false
      required:
        [
          reads,
          sales_website,
          leads_website,
          leads_instant_form,
          media_upload,
          proposal_creation,
        ]
      properties:
        reads: { $ref: "#/components/schemas/CapabilityCheck" }
        sales_website: { $ref: "#/components/schemas/CapabilityCheck" }
        leads_website: { $ref: "#/components/schemas/CapabilityCheck" }
        leads_instant_form: { $ref: "#/components/schemas/CapabilityCheck" }
        media_upload: { $ref: "#/components/schemas/CapabilityCheck" }
        proposal_creation: { $ref: "#/components/schemas/CapabilityCheck" }
    CapabilityAccount:
      type: object
      additionalProperties: false
      required: [currency, timezone]
      properties:
        currency:
          oneOf: [{ $ref: "#/components/schemas/Currency" }, { type: "null" }]
          description: Null only when unavailable and accompanied by gap code account_currency_unavailable; never fabricated.
        timezone:
          type: [string, "null"]
          minLength: 1
          description: Meta-configured IANA timezone, or null with gap code account_timezone_unavailable; never fabricated.
    CapabilityAssetType:
      type: string
      enum: [page, pixel, web_dataset, lead_form, instagram_account]
    PageAsset:
      type: object
      additionalProperties: false
      required: [asset_type, page_id, name]
      properties:
        asset_type: { const: page }
        page_id: { $ref: "#/components/schemas/Identifier" }
        name: { type: string, minLength: 1 }
    PixelAsset:
      type: object
      additionalProperties: false
      required: [asset_type, pixel_id, name]
      properties:
        asset_type: { const: pixel }
        pixel_id: { $ref: "#/components/schemas/Identifier" }
        name: { type: string, minLength: 1 }
    WebDatasetAsset:
      type: object
      additionalProperties: false
      required: [asset_type, pixel_id, name]
      properties:
        asset_type: { const: web_dataset }
        pixel_id: { $ref: "#/components/schemas/Identifier" }
        name: { type: string, minLength: 1 }
      description: Meta web datasets use pixel_id semantics in this API.
    LeadFormAsset:
      type: object
      additionalProperties: false
      required: [asset_type, lead_gen_form_id, page_id, name, published, usable]
      properties:
        asset_type: { const: lead_form }
        lead_gen_form_id: { $ref: "#/components/schemas/Identifier" }
        page_id: { $ref: "#/components/schemas/Identifier" }
        name: { type: string, minLength: 1 }
        published: { const: true }
        usable: { const: true }
...
    CapabilityAsset:
      oneOf:
        - $ref: "#/components/schemas/PageAsset"
        - $ref: "#/components/schemas/PixelAsset"
        - $ref: "#/components/schemas/WebDatasetAsset"
        - $ref: "#/components/schemas/LeadFormAsset"
        - $ref: "#/components/schemas/InstagramAccountAsset"
      discriminator:
        propertyName: asset_type
        mapping:
          page: "#/components/schemas/PageAsset"
          pixel: "#/components/schemas/PixelAsset"
          web_dataset: "#/components/schemas/WebDatasetAsset"
          lead_form: "#/components/schemas/LeadFormAsset"
          instagram_account: "#/components/schemas/InstagramAccountAsset"
    CapabilitiesResponse:
      type: object
      additionalProperties: false
      required:
        [
          request_id,
          scope,
          integration,
          account,
          supported_campaign_kinds,
          assets,
          page,
          capabilities,
          gaps,
        ]
      properties:
        request_id: { $ref: "#/components/schemas/RequestId" }
        scope: { $ref: "#/components/schemas/ResolvedScope" }
        integration: { $ref: "#/components/schemas/IntegrationStatus" }
        account: { $ref: "#/components/schemas/CapabilityAccount" }
        supported_campaign_kinds:
          type: array
          minItems: 3
          maxItems: 3
          uniqueItems: true
          items: { $ref: "#/components/schemas/CampaignKind" }
          description: Exactly SALES_WEBSITE, LEADS_WEBSITE, and LEADS_INSTANT_FORM.
        assets:
          type: array
          items: { $ref: "#/components/schemas/CapabilityAsset" }
          description: |
            One deterministic page of accessible assets only. Order is stable by
            asset_type then opaque asset ID. IDs confer no authority.
        page: { $ref: "#/components/schemas/PageInfo" }
        capabilities: { $ref: "#/components/schemas/CapabilityMatrix" }
        gaps:
          type: array
          uniqueItems: true
          items: { $ref: "#/components/schemas/Diagnostic" }
          description: |
            Actionable reasons and remediation only; codes are unique and never contain
            tokens, secrets, or authorization material. Every unavailable capability
            diagnostic code matches exactly one entry here. Null currency and timezone
            require account_currency_unavailable and account_timezone_unavailable respectively.
      allOf:
        - if:
            properties:
              account:
                type: object
                required: [currency]
                properties: { currency: { type: "null" } }
          then:
            properties:
              gaps:
                contains:
                  {
                    type: object,
                    required: [code],
                    properties:
                      { code: { const: account_currency_unavailable } },
                  }
        - if:
            properties:
              account:
                type: object
                required: [timezone]
                properties: { timezone: { type: "null" } }
          then:
            properties:
              gaps:
                contains:
                  {
                    type: object,
                    required: [code],
                    properties:
                      { code: { const: account_timezone_unavailable } },
                  }
...
    AttachmentMetadata:
      type: object
      additionalProperties: false
      required:
        [source, attachment_id, original_filename, declared_content_type]
      properties:
        source: { const: openclaw_chat_attachment }
        attachment_id: { $ref: "#/components/schemas/Identifier" }
        original_filename:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '^[^/\\\x00-\x1F\x7F]+$'
          description: Basename display metadata only; slash, backslash, and control characters are rejected. Never interpreted as a path.
        declared_content_type:
          { type: string, enum: [image/jpeg, image/png, video/mp4] }
        alt_text: { type: string, maxLength: 1000 }
    MediaUpload:
      type: object
      additionalProperties: false
      required: [client_id, ad_account_id, file, attachment]
      properties:
        client_id: { $ref: "#/components/schemas/Identifier" }
        ad_account_id: { $ref: "#/components/schemas/Identifier" }
        file: { type: string, format: binary }
        attachment: { $ref: "#/components/schemas/AttachmentMetadata" }
    StagedMedia:
      type: object
      additionalProperties: false
      required:
        [
          media_id,
          sha256,
          status,
          media_type,
          content_type,
          size_bytes,
          created_at,
          expires_at,
        ]
      properties:
        media_id: { $ref: "#/components/schemas/Identifier" }
        sha256: { type: string, pattern: "^[a-f0-9]{64}$" }
        status:
          { type: string, enum: [staged, bound, consumed, expired, invalid] }
        media_type: { type: string, enum: [image, video] }
        content_type: { type: string, enum: [image/jpeg, image/png, video/mp4] }
        size_bytes: { type: integer, minimum: 1 }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        operation_id:
          oneOf: [{ $ref: "#/components/schemas/Identifier" }, { type: "null" }]
    MediaResponse:
      type: object
      additionalProperties: false
      required: [request_id, scope, media]
      properties:
        request_id: { $ref: "#/components/schemas/RequestId" }
        scope: { $ref: "#/components/schemas/ResolvedScope" }
        media: { $ref: "#/components/schemas/StagedMedia" }
...
    CapabilityProblem400:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/CapabilityProblemBase"
        - {
            type: object,
            properties:
              {
                status: { const: 400 },
                code:
                  { type: string, enum: [validation_error, cursor_mismatch] },
              },
          }
    CapabilityProblem401:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/CapabilityProblemBase"
        - {
            type: object,
            properties:
              { status: { const: 401 }, code: { const: unauthorized } },
          }
    CapabilityProblem403:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/CapabilityProblemBase"
        - {
            type: object,
            properties: { status: { const: 403 }, code: { const: forbidden } },
          }
    CapabilityProblem409:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/CapabilityProblemBase"
        - type: object
          required: [supplied_scope]
          properties:
            status: { const: 409 }
            code: { const: client_account_mismatch }
            supplied_scope:
              { $ref: "#/components/schemas/CapabilityRequiredSuppliedScope" }
    Problem400:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/ProblemBase"
        - {
            type: object,
            properties:
              {
                status: { const: 400 },
                code:
                  { type: string, enum: [validation_error, cursor_mismatch] },
              },
          }
    Problem401:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/ProblemBase"
        - {
            type: object,
            properties:
              { status: { const: 401 }, code: { const: unauthorized } },
          }
    Problem403:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/ProblemBase"
        - {
            type: object,
            properties: { status: { const: 403 }, code: { const: forbidden } },
          }
    Problem404:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/ProblemBase"
        - {
            type: object,
            properties: { status: { const: 404 }, code: { const: not_found } },
          }
    Problem409Scope:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/ProblemBase"
        - {
            type: object,
            properties:
              {
                status: { const: 409 },
                code: { const: client_account_mismatch },
              },
          }
...
    Problem413:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/ProblemBase"
        - {
            type: object,
            properties:
              { status: { const: 413 }, code: { const: media_too_large } },
          }
...
    Problem422Media:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/ProblemBase"
        - {
            type: object,
            properties:
              { status: { const: 422 }, code: { const: media_invalid } },
          }
...
    Problem429:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/ProblemBase"
        - {
            type: object,
            properties:
              { status: { const: 429 }, code: { const: rate_limited } },
          }
    Problem500:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/ProblemBase"
        - {
            type: object,
            properties:
              { status: { const: 500 }, code: { const: internal_error } },
          }
    Problem502:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/ProblemBase"
        - {
            type: object,
            properties: { status: { const: 502 }, code: { const: meta_error } },
          }
security:
  - ServiceBearer: []